v0.7.4 [algorithm] type fixs...........................

+++ src/algorithms/FastDijkstra.cc
@@ -11,25 +11,24 @@
 
 using namespace std;
 
-typedef pair<int, int> PII;
-typedef vector<pair<int, int>> VPII;
-typedef vector<int> VI;
-
-VI Dijkstra(VPII edges, int s) {
+typedef pair<uint64_t, uint64_t> PII;
+typedef vector<pair<uint64_t, uint64_t>> VPII;
+typedef vector<uint64_t> VI;
+VI Dijkstra(VPII edges, uint64_t s) {
   // use priority queue in which top element has the "smallest" priority
   priority_queue<PII, VPII, greater<PII>> Q;
   VI dist(edges.size(), INFINITY), dad(edges.size(), -1);
-  Q.push(pair<int, int>(0, s));
+  Q.push(pair<uint64_t, uint64_t>(0, s));
   dist[s] = 0;
   while(!Q.empty()){
     PII p = Q.top();
     Q.pop();
-    int here = p.second;
-    /*for(VPII::iterator it=edges[here].begin(); it!=edges[here].end(); it++){
+    uint64_t here = p.second;
+/*    for(VPII::iterator it=edges[here].begin(); it!=edges[here].end(); it++){
       if(dist[here] + it->first < dist[it->second]){
         dist[it->second] = dist[here] + it->first;
         dad[it->second] = here;
-        Q.push (pair<int, int>(dist[it->second], it->second));
+        Q.push (pair<uint64_t, uint64_t>(dist[it->second], it->second));
       }
     }*/
   }
